Remove OS specific toolboxes
It is doesn't seem possible to provide multiple toolboxes in a single release
in a robust way with the File Exchange. The File Exchange automatically grabs a
new toolbox from github when a release is pushed but it does not give a way to
specify which toolbox it should get. Ideally the File Exchange would be able to
handle OS specific distributions but at the moment it cannot.