Use older GCC/G++ compiler for compatibility with MATLAB libs
Instead of trying to ship our own newer libs. I believe that was not working
because MATLAB already dynamically loads its own libs at start up, and multiple
versions of the same library cannot be loaded at the same time because that
would lead to multiple definitions of the same symbols. This means the RPATH
cannot ensure the libraries we won't will be loaded.

Actually fixes #2.