Add reindex membership function
Many igraph functions require membership vectors to be sequential. This
provides a method to repack the community labels to be between 1 and
n_unique_communities. Unlike the upstream igraph reindex membership function
this handles cases where the largest community number is larger than the number
of nodes in the membership vector. This also runs on all membership vectors in
a matrix (where each membership vector is one row).