Don't free stack on user interruption
Since an error should be triggered in response to a user interruption sooner or
later, and the error will call IGRAPH_FINALLY_FREE, it does not need to be
called in the interruption handler. And it's possible some of the memory that
will be cleaned up when freeing the stack will be needed as part of the clean
up between detecting the interruption and calling the error.